change cli to use both commands, add utils for common comands
